Selector Fecha

+++ plantilla.py
@@ -4,9 +4,10 @@ import tkinter as tk
 import tkinter.ttk as ttk
 from tkinter import messagebox as mssg
 import sqlite3
+from datetime import date, datetime
 
 class Participantes:
-    # nombre de la base de datos  y ruta 
+    # nombre de la base de datos y ruta 
     path = r'Resources'
     db_name = path + r'/Participantes.db'
     program_icon = path + r'/ico_registro.ico'
@@ -17,8 +18,11 @@ class Participantes:
         'entry': '#FFFFFF'
     }
 
+    # Variables usadas para guardar información digitada por el usuario
     cod_departamento = None
     cod_ciudad = None
+    sel_fecha = None #si hay una fecha seleccionada esto será un objeto date
+
     def __init__(self, master=None):
         # Top Level - Ventana Principal
         self.win = tk.Tk() if master is None else tk.Toplevel()
@@ -110,9 +114,8 @@ class Participantes:
         self.entryCiudad.grid(column='0', row='0', sticky='w')
 
         # Botón de selección
-        self.btnSeleccionarCiudad = ttk.Button(self.frmCiudad, text='...', width=2)
+        self.btnSeleccionarCiudad = ttk.Button(self.frmCiudad, text='...', width=2, command=self.crear_Selector_Ciudad)
         self.btnSeleccionarCiudad.grid(column='1', row='0', padx=5)
-        self.btnSeleccionarCiudad.bind('<1>', self.crear_Selector_Ciudad)
         
         #Label Direccion
         self.lblDireccion = ttk.Label(self.lblfrm_Datos)
@@ -158,14 +161,22 @@ class Participantes:
         self.lblFecha.configure(config_lbl_Datos)
         self.lblFecha.configure(text='Fecha', width='12')
         self.lblFecha.grid(column='0', padx='5', row='6', sticky='w')
+
+        #Frame Entry Fecha (background coincide con lblfrmDatos)
+        self.frmFecha = tk.Frame(self.lblfrm_Datos, background=self.color_palette['lblfrm_datos'])
+        self.frmFecha.grid(column='1', row='6', sticky='w')
         
         #Entry Fecha
         self.entryFechaText = tk.StringVar()
-        self.entryFechaText.trace_add('write', self.valida_Fecha)
-        self.entryFecha = tk.Entry(self.lblfrm_Datos, textvariable=self.entryFechaText)
+        self.entryFechaText.set('[Seleccionar]')
+        self.entryFecha = tk.Entry(self.frmFecha, textvariable=self.entryFechaText)
         self.entryFecha.configure(config_entry_Datos)
-        self.entryFecha.configure(width=30)
-        self.entryFecha.grid(column='1', row='6', sticky='w')
+        self.entryFecha.configure(width=25)
+        self.entryFecha.grid(column='0', row='0', sticky='w')
+
+        # Botón de selección
+        self.btnSeleccionarCiudad = ttk.Button(self.frmFecha, text='...', width=2, command=self.crear_Selector_Fecha)
+        self.btnSeleccionarCiudad.grid(column='1', row='0', padx=5)
         
         #Frame Botones (background coincide con la ventana)
         self.frmBotones = tk.Frame(self.win, background=self.color_palette['window_bg'])
@@ -260,7 +271,20 @@ class Participantes:
         db_rows = self.run_Query(query)
         # Insertando los datos de la BD en la tabla de la pantalla
         for row in db_rows:
-            self.treeDatos.insert('',0, text = row[0], values = [row[1],row[2],row[3],row[4],row[5],row[6]])
+            ciudad = row[2]
+            if ciudad is None:
+                ciudad = ''
+            fecha = row[6]
+            if fecha is None:
+                fecha = ''
+            else:
+                try:
+                    # Si hay una fecha valida la convierte a DD/MM/AAAA
+                    fecha = datetime.strptime(fecha, '%Y-%m-%d').strftime('%d/%m/%Y')
+                except ValueError:
+                    # Si hay una fecha invalida pone el texto como 'Invalida'abs
+                    fecha = 'Invalida'
+            self.treeDatos.insert('',0, text = row[0], values = [row[1],ciudad,row[3],row[4],row[5],fecha])
 
     def valida_Grabar(self):
         '''Valida que el Id no esté vacio, devuelve True si ok'''
@@ -282,27 +306,6 @@ class Participantes:
         #Pone en la StringVar el texto filtrado recortado a los primeros 15 caracteres
         self.entryIdText.set(filtered_text[0:15])
 
-    def valida_Fecha(self, var, index, mode):
-        text = self.entryFechaText.get()
-        cursor_pos = self.entryFecha.index(tk.INSERT)
-        filtered_text = ''
-        for char in text:
-            if char in {'1', '2', '3', '4', '5', '6', '7', '8', '9', '0'}:
-                filtered_text += char
-        filtered_text = filtered_text[0:8]
-
-        if len(filtered_text) >= 4:
-            formatted_text = filtered_text[0:2] + '/' + filtered_text [2:4] + '/' + filtered_text[4:]
-        elif len(filtered_text) >= 2:
-            formatted_text = filtered_text[0:2] + '/' + filtered_text [2:]
-        else:
-            formatted_text = filtered_text
-
-        if len(filtered_text) in {2, 4}:
-            cursor_pos += 1
-
-        self.entryFechaText.set(formatted_text)
-        self.win.after(1, self.reset_cursor, self.entryFecha, cursor_pos)
     
     '''Función utilizada para el boón editar, trae desde el treeview el participante seleccionado
         y lo carga en los entry '''
@@ -311,13 +314,12 @@ class Participantes:
         #   Nuevo
         self.entryIdText.set(self.treeDatos.item(self.treeDatos.selection())['text']) #Aca, todo lo que habia se reemplaza por el texto nuevom si coloco insert, se acumulan
         self.entryId.configure(state = 'readonly')#Deja id bloqueado para que no lo puedan editar
-        #Carga los demás entries
+        #Carga los demás entries (para ciudad esto es meramente el texto mostrado)
         self.entryNombreText.set(self.treeDatos.item(self.treeDatos.selection())['values'][0])
         self.entryCiudadText.set(self.treeDatos.item(self.treeDatos.selection())['values'][1])
         self.entryDireccionText.set(self.treeDatos.item(self.treeDatos.selection())['values'][2])
         self.entryCelularText.set(self.treeDatos.item(self.treeDatos.selection())['values'][3])
         self.entryEntidadText.set(self.treeDatos.item(self.treeDatos.selection())['values'][4])
-        self.entryFechaText.set(self.treeDatos.item(self.treeDatos.selection())['values'][5])
 
         # Carga los códigos del departamento y de la ciudad
         query = 'SELECT Id_Departamento, Id_Ciudad FROM t_participantes WHERE Id = ?'
@@ -325,6 +327,22 @@ class Participantes:
         codigos = self.run_Query(query, parametros).fetchone()
         self.cod_departamento = codigos[0]
         self.cod_ciudad = codigos[1]
+
+        # Si la ciudad no está definida, dejar el campo "vacio"
+        if self.cod_ciudad is None:
+            self.entryCiudadText.set('[Seleccionar]')
+
+        # Carga la fecha si no está vacia
+        text_fecha = self.treeDatos.item(self.treeDatos.selection())['values'][5]
+        try:
+            # Si hay una fecha valida, la guarda en self.fecha y pone el texto del entry en la fecha correspondiente (recordar que en el treeview ya está en DD/MM/AAAA)
+            self.sel_fecha = datetime.strptime(text_fecha, '%d/%m/%Y').date()
+            self.entryFechaText.set(text_fecha)
+        except ValueError:
+            # Sino, pone el campo "vacio"
+            self.fecha = None
+            self.entryFechaText.set('[Seleccionar]')
+
         
 
     def limpia_Campos(self):
@@ -336,9 +354,10 @@ class Participantes:
         self.entryDireccionText.set('')
         self.entryCelularText.set('')
         self.entryEntidadText.set('')
-        self.entryFechaText.set('')
+        self.entryFechaText.set('[Seleccionar]')
         self.cod_departamento = None
         self.cod_ciudad = None
+        self.sel_fecha = None
     
 
     def get_Departamentos(self):
@@ -348,13 +367,9 @@ class Participantes:
     def get_Ciudades_Por_Departamento(self, departamento):
         query = 'SELECT Nombre_Ciudad FROM t_ciudades WHERE Nombre_Departamento = ? ORDER BY Nombre_Ciudad'
         parametros = (departamento,)
-        try:
-            ciudades = [row[0] for row in self.run_Query(query, parametros)]
-            return ciudades
-        except sqlite3.OperationalError as e:
-            return []
+        return [row[0] for row in self.run_Query(query, parametros)]
         
-    def crear_Selector_Ciudad(self, event=None):
+    def crear_Selector_Ciudad(self):
         pre_cod_departamento = None
         pre_cod_departamento_text = None
         pre_cod_ciudad = None
@@ -439,10 +454,7 @@ class Participantes:
     
         def seleccionar_Departamento(event):
             # permite modificar las variables definidas en crear_Selector_Ciudad
-            nonlocal pre_cod_departamento 
-            nonlocal pre_cod_departamento_text
-            nonlocal pre_cod_ciudad
-            nonlocal pre_cod_ciudad_text
+            nonlocal pre_cod_departamento, pre_cod_departamento_text, pre_cod_ciudad, pre_cod_ciudad_text
             seleccion_departamento = listboxDepartamentos.curselection()
             if seleccion_departamento:
                 # Si se selecciona otro departamento se deselecciona la ciudad
@@ -463,8 +475,7 @@ class Participantes:
                 
     
         def seleccionar_Ciudad(event):
-            nonlocal pre_cod_ciudad
-            nonlocal pre_cod_ciudad_text
+            nonlocal pre_cod_ciudad, pre_cod_ciudad_text
             seleccion_ciudad = listboxCiudades.curselection()
 
             if seleccion_ciudad:
@@ -492,22 +503,140 @@ class Participantes:
         btnCancelar.configure(command=ventana.destroy)
         btnSeleccionar.configure(command=confirmar_Seleccion)
 
+    def num_Dias_Mes(self, anio, mes):
+        '''Devuelve la cantidad de dias entre el primer dia del mes y el primer dia del mes siguiente'''
+        sig_mes = mes + 1
+        sig_anio = anio
+        
+        if mes == 12:
+            sig_mes = 1
+            sig_anio = sig_anio + 1
+
+        time_difference = date(sig_anio, sig_mes, 1) - date(anio, mes, 1)
+        print(time_difference.days)
+        return time_difference.days
+
+    def crear_Selector_Fecha(self):
+        current_date = date.today()
+        current_day = current_date.day
+        current_month = current_date.month
+        current_year = current_date.year
+
+        # Colocar por defecto el día actual o el seleccionado
+        if self.sel_fecha is not None:
+            sel_dia = self.sel_fecha.day
+            sel_mes = self.sel_fecha.month
+            sel_anio = self.sel_fecha.year
+        else:
+            sel_dia = current_day
+            sel_mes = current_month
+            sel_anio = current_year
+
+        ventana = tk.Toplevel(self.win)
+        ventana.title('Seleccionar Fecha')
+        ventana.transient(self.win)
+
+        lblSeleccion = ttk.Label(ventana, anchor='center', text='Seleccione una fecha')
+        lblSeleccion.pack(side='top', fill='x', padx=5, pady=5)
+
+        frmBotones = tk.Frame(ventana)
+        frmBotones.pack(side='bottom', pady=5)
+
+        btnSeleccionar = ttk.Button(frmBotones, text='Seleccionar')
+        btnSeleccionar.grid(row=0, column=0, padx=5)
+
+        btnCancelar = ttk.Button(frmBotones, text='Cancelar')
+        btnCancelar.grid(row=0, column=1, padx=5)
+
+
+        # Crear frame con selectores de fecha
+        frmFecha = tk.Frame(ventana)
+        frmFecha.configure(height=50, pady=5, padx=5)
+        frmFecha.pack(side='bottom', anchor='s')
+
+        lblDia = tk.Label(frmFecha, text='Día')
+        lblDia.grid(row=0, column=0, padx=5, pady=3)
+        comboDia = ttk.Combobox(frmFecha, state='readonly', width=5)
+        comboDia.grid(row=1, column=0, pady=5, padx=7)
+        
+
+        lblMes = tk.Label(frmFecha, text='Mes')
+        lblMes.grid(row=0, column=1, padx=5, pady=3)
+        comboMes = ttk.Combobox(frmFecha, state='readonly', width=7)
+        comboMes.grid(row=1, column=1, pady=5, padx=7)
+        
+
+        lblAnio = tk.Label(frmFecha, text='Año')
+        lblAnio.grid(row=0, column=2, padx=5, pady=3)
+        comboAnio = ttk.Combobox(frmFecha, state='readonly', width=7)
+        comboAnio.grid(row=1, column=2, pady=5, padx=7)
+        
+        def rellenar_Selectores_Fecha():
+            num_dias_mes = self.num_Dias_Mes(sel_anio, sel_mes)
+            
+            comboAnio['values'] = tuple(range(1900, 2100 + 1)) #tupla del 1900 al 2100
+            comboMes['values'] = ('Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio', 'Agosto', 'Septiembre', 'Octubre', 'Noviembre', 'Diciembre')
+            comboDia['values'] = tuple(range(1, num_dias_mes + 1)) #tupla desde al 1 hasta el último dia del mes
+
+        # Inicializar en el día actual/seleccionado (el index será el número correspondiente menos el primer número posible)
+        rellenar_Selectores_Fecha()
+        comboDia.set(sel_dia)
+        comboMes.current(sel_mes - 1)
+        comboAnio.set(sel_anio)
+
+        def actualizar_Seleccion(event = None):
+            nonlocal sel_anio, sel_mes, sel_dia
+            sel_anio = int(comboAnio.get())
+            sel_mes = comboMes.current() + 1
+            sel_dia = int(comboDia.get())
+
+            num_dias_mes = self.num_Dias_Mes(sel_anio, sel_mes)
+            #Si el dia seleccionado es mayor al último día del mes, forzamos la selección al último día del mes
+            if sel_dia > num_dias_mes:
+                sel_dia = num_dias_mes 
+                comboDia.set(num_dias_mes)
+            rellenar_Selectores_Fecha()
+
+        def confirmar_Seleccion():
+            if date(sel_anio, sel_mes, sel_dia) > current_date:
+                mssg.showerror('¡ Atención !', 'No se puede seleccionar una fecha posterior a hoy')
+            elif date(sel_anio, sel_mes, sel_dia) < date(current_year, 1, 1):
+                mssg.showerror('¡ Atención !', f'No se puede seleccionar una fecha antes de {current_year}')
+            else:
+                self.sel_fecha = datetime(sel_anio, sel_mes, sel_dia)
+                self.entryFechaText.set(self.sel_fecha.strftime('%d/%m/%Y'))
+                ventana.destroy()
+
+
+        comboDia.bind('<<ComboboxSelected>>', actualizar_Seleccion)
+        comboMes.bind('<<ComboboxSelected>>', actualizar_Seleccion)
+        comboAnio.bind('<<ComboboxSelected>>', actualizar_Seleccion)
+        btnSeleccionar.configure(command=confirmar_Seleccion)
+        btnCancelar.configure(command=ventana.destroy)
+
+
 
     def adiciona_Registro(self):
         '''Adiciona un producto a la BD si la validación es True'''
+        # Convierte la fecha de datetime a AAAA-MM-DD (así lo maneja sqlite)
+        if self.sel_fecha is not None:
+            fecha_sql = self.sel_fecha.strftime('%Y-%m-%d')
+        else:
+            fecha_sql = None
+        
         if self.actualiza:
             self.actualiza = None #REVISION
             self.entryId.configure(state = 'readonly')
             query = 'UPDATE t_participantes SET Nombre = ?, Direccion = ?, Celular = ?, Entidad = ?, Fecha = ?, Id_Departamento = ?, Id_Ciudad = ? WHERE Id = ?'
             parametros = (self.entryNombre.get(), self.entryDireccion.get(), self.entryCelular.get(),
-                          self.entryEntidad.get(), self.entryFecha.get(), self.cod_departamento,
+                          self.entryEntidad.get(), fecha_sql, self.cod_departamento,
                           self.cod_ciudad, self.entryId.get())
             self.run_Query(query, parametros)
             mssg.showinfo('Ok', 'Registro actualizado con éxito')
         else:
             query = 'INSERT INTO t_participantes(Id, Nombre, Direccion, Celular, Entidad, Fecha, Id_Departamento, Id_Ciudad) VALUES(?, ?, ?, ?, ?, ?, ?, ?)'
             parametros = (self.entryId.get(), self.entryNombre.get(), self.entryDireccion.get(),
-                          self.entryCelular.get(), self.entryEntidad.get(), self.entryFecha.get(),
+                          self.entryCelular.get(), self.entryEntidad.get(), fecha_sql,
                           self.cod_departamento, self.cod_ciudad)
             if self.valida_Grabar():
                 self.run_Query(query, parametros)
@@ -519,38 +648,29 @@ class Participantes:
         self.lee_tablaTreeView()
 
     def edita_tablaTreeView(self):
-        #if len(self.treeDatos.selection()) > 1:
-            #print(len(self.treeDatos.selection()))
-            #raise mssg.showerror('¡ Atención !', 'Solo se puede editar un item a la vez')
-        #else:
-            try:
-                if len(self.treeDatos.selection()) > 1:
-                    temp = True
-                    raise IndexError
-                else:
-                    # Carga los campos desde la tabla TreeView
-                    self.treeDatos.item(self.treeDatos.selection())['text']
-                    self.limpia_Campos()
-                    self.actualiza = True # Esta variable controla la actualización
-                    self.carga_Datos()
-            except IndexError as error:
-                self.actualiza = None
-                if temp:
-                    mssg.showerror('¡ Atención !', 'Solo puede seleccionar un ítem a editar')
-                else:
-                    mssg.showerror('¡ Atención !','Por favor seleccione un ítem de la tabla')
-                return
+        if len(self.treeDatos.selection()) == 0:
+            mssg.showerror('¡ Atención !', 'Por favor seleccione un ítem de la tabla')
+        elif len(self.treeDatos.selection()) > 1:
+            mssg.showerror('¡ Atención !', 'Solo puede seleccionar un ítem a editar')
+        else:
+            self.treeDatos.item(self.treeDatos.selection())['text']
+            self.limpia_Campos()
+            self.actualiza = True # Esta variable controla la actualización
+            self.carga_Datos()
         
     def elimina_Registro(self):
         if not self.treeDatos.selection(): #Verifica si la selección de la tabla esta vacía, osea, no ha seleccionado nada
-            mssg.showinfo('¡ Atención !', 'Por favor seleciones los items a eliminar de la tabla')
-        else:
+            mssg.showerror('¡ Atención !', 'Por favor seleccione los items a eliminar de la tabla')
+            return
+        
+        confirmation_msg = mssg.askyesno('', '¿Desea borrar los datos seleccionados?')
+        if confirmation_msg == True:
             query = ('DELETE FROM t_participantes WHERE Id = ?')
             for param in self.treeDatos.selection(): #Bucle para que pueda borrar cada selección
                 parametros = (self.treeDatos.item(param)['text'], ) #Una selección que hace el usuario
                 self.run_Query(query, parametros)
             
-            mssg.showerror('Eliminado', 'Los registros seleccionados fueron eliminados')
+            mssg.showinfo('Eliminado', 'Los registros seleccionados fueron eliminados')
             self.lee_tablaTreeView() #Carga la tabla al treeview actualizada
